% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.file.R
\name{crawl.file}
\alias{crawl.file}
\title{A crawl.file Function}
\usage{
crawl.file(destfile, url, fake.header = NULL, verbose = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
