% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.washing.R
\name{data.washing}
\alias{data.washing}
\title{A data_Washing Function}
\usage{
data.washing()
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
