% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Washing.R
\name{data_Washing}
\alias{data_Washing}
\title{A data_Washing Function}
\usage{
data_Washing()
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
