% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.MGR.R
\name{dataset.MGR}
\alias{dataset.MGR}
\title{A dataset.MGR Function}
\usage{
dataset.MGR(
  dataset.name = NULL,
  group = c("stock", "data"),
  request = "conf",
  force.update = FALSE,
  file.extension = ".csv",
  header = TRUE,
  to.exts = FALSE,
  online = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
