% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbg.R
\name{dbg}
\alias{dbg}
\title{A dbg Function}
\usage{
dbg(x = NULL, type = "print", name, pause = FALSE, debug = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
