% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downbreak.R
\name{downbreak}
\alias{downbreak}
\title{A downbreak Function}
\usage{
downbreak(Line, RefLine)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
