% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef.fund.Allocation.R
\name{ef.fund.Allocation}
\alias{ef.fund.Allocation}
\title{fund.Allocation Function}
\usage{
ef.fund.Allocation(r_set, goal_return = 0.000701, type = 1)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
