% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.of.stockData.R
\name{generator.of.stockData}
\alias{generator.of.stockData}
\title{A generator.of.stockDat Function}
\usage{
generator.of.stockData(
  tranSet.period = NULL,
  data.path = NULL,
  list.path = NULL,
  group = NULL,
  auto.detect = TRUE
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
