% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.conf.R
\name{get.conf}
\alias{get.conf}
\title{A multi sort Function}
\usage{
get.conf(
  name,
  env.name = "env",
  default.conf = NULL,
  splite.char = FALSE,
  splite.num = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
