% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_get.script.name.R
\name{get.script.name}
\alias{get.script.name}
\title{A get.script.name Function}
\usage{
get.script.name()
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
