% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sysinfo.R
\name{get.sysinfo}
\alias{get.sysinfo}
\title{A get sysinfo Function}
\usage{
get.sysinfo(x = 0, path = NULL)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
