% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.users.input.R
\name{get.users.input}
\alias{get.users.input}
\title{A get.users.input Function}
\usage{
get.users.input(prompt = NULL, index = NULL)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
