% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdj.Indicator.R
\name{kdj.Indicator}
\alias{kdj.Indicator}
\title{A kdj.Indicator Function}
\usage{
kdj.Indicator(data_Import, n = 9)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
