% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_backtest.R
\name{m_backtest}
\alias{m_backtest}
\title{A backtestFunction}
\usage{
m_backtest(ret, object2)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
