% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_check.code.R
\name{m_check.code}
\alias{m_check.code}
\title{A m_check.code Function}
\usage{
m_check.code(string, num = 4)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
