% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_dupli.vector.R
\name{m_dupli.vector}
\alias{m_dupli.vector}
\title{A m_dupli.vector Function}
\usage{
m_dupli.vector(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
