% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_env.R
\name{m_env}
\alias{m_env}
\title{A enchaned env Function}
\usage{
m_env(name = NULL, value = NULL, mode = "init_list", dataset = "env")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
