% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_get.data.R
\name{m_get.data}
\alias{m_get.data}
\title{A m_get.data Function}
\usage{
m_get.data(list, header = TRUE, data.col = 1)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to get data from list.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
