% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_gsub.R
\name{m_gsub}
\alias{m_gsub}
\title{A multi gsub Function}
\usage{
m_gsub(be_replace, patten, v, fixed = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
