% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_indicator.filter.R
\name{m_indicator.filter}
\alias{m_indicator.filter}
\title{A m_indicator.filter Function}
\usage{
m_indicator.filter(ind, level = 1)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
