% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_lag.R
\name{m_lag}
\alias{m_lag}
\title{Am_lag Function}
\usage{
m_lag(x, lag.num = 1, shift = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
