% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_msg.R
\name{m_msg}
\alias{m_msg}
\title{A message Function}
\usage{
m_msg(info, order = "\%X", action = "msg")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
