% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_page2table.R
\name{m_page2table}
\alias{m_page2table}
\title{A page2table Function}
\usage{
m_page2table(
  page.url,
  page.encoding = "UTF-8",
  export.file = NULL,
  css.selector = NULL
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
