% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_paste.R
\name{m_paste}
\alias{m_paste}
\title{A multi paste Function}
\usage{
m_paste(x, op = "")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
