% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_performance.test.R
\name{m_performance.test}
\alias{m_performance.test}
\title{A muti_backtestt Function}
\usage{
m_performance.test(
  selected.list = NULL,
  testSet_period = NULL,
  tranSet_period = NULL,
  trading.straregy_type = c(2, 80, 20, 80, 20),
  ef_goal_return = 0.001,
  enable_reandom_stocks = FALSE,
  enable_fund_allocation = TRUE,
  ef_simulation_num = 50,
  enable_ef_simulation.record = TRUE,
  memo = "",
  job.id = NULL
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
