% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_rebind.xts.R
\name{m_rebind.xts}
\alias{m_rebind.xts}
\title{A m_rebind.xtsFunction}
\usage{
m_rebind.xts(x, v, byrow = TRUE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
