% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_regist.task.status.R
\name{m_regist.task.status}
\alias{m_regist.task.status}
\title{A get.script.name Function}
\usage{
m_regist.task.status(task.value = NULL, mode = "w")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
