% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_regressor5.R
\name{m_regressor.fit}
\alias{m_regressor.fit}
\title{A regressor.fit Function}
\usage{
m_regressor.fit(
  stock.code,
  tranSet.period,
  testSet.period,
  data.path,
  retma.data.path,
  lag.num = 20
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
