% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_remix.data.R
\name{m_remix.data}
\alias{m_remix.data}
\title{A m_remix.data Function}
\usage{
m_remix.data(data = NULL, mode = NULL)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
