% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_series.R
\name{m_series}
\alias{m_series}
\title{A m_series Function}
\usage{
m_series(type = "Fibo", n = 15)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
