% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_smaCal.R
\name{m_smaCal}
\alias{m_smaCal}
\title{A m_std.data Function}
\usage{
m_smaCal(ts, n = NULL, increase = TRUE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
