% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sort.R
\name{m_sort}
\alias{m_sort}
\title{A multi sort Function}
\usage{
m_sort(x, key = NULL, decreasing = FALSE, reindex = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
