% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_std.data.R
\name{m_std.data}
\alias{m_std.data}
\title{A m_std.data Function}
\usage{
m_std.data(x, restore = NULL)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
