% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_time.invested.R
\name{m_time.invested}
\alias{m_time.invested}
\title{A m_time.invested Function}
\usage{
m_time.invested(data = NULL, ma.VALUE = c(3, 5, 20))
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
