% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.data_ByFile.R
\name{merge.data_ByFile}
\alias{merge.data_ByFile}
\title{A for merge.data_ByFile  Function}
\usage{
\method{merge}{data_ByFile}(name, period = NULL, order = NULL, debugger = FALSE, header = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
