% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muti_backtest.R
\name{muti_backtest}
\alias{muti_backtest}
\title{A muti_backtestt Function}
\usage{
muti_backtest()
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
