% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_misc.R
\name{na.filter}
\alias{na.filter}
\title{A MISC Function}
\usage{
na.filter(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
