% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurons.filter.R
\name{neurons.filter}
\alias{neurons.filter}
\title{A neurons.filter Function}
\usage{
neurons.filter(x, type = "pureline")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
