% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation.indicator.R
\name{oscillation.indicator}
\alias{oscillation.indicator}
\title{A surge.indicator Function}
\usage{
oscillation.indicator(
  Clprs = Clprs,
  trigger.action = c("brown", "orange"),
  trigger.mat = c(-0.005, 0.005, -0.01, 0.01, -0.015, 0.015, -0.025, 0.025, -0.04,
    0.04),
  record.path = NULL,
  auto.detect = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
