% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page2table.R
\name{page2table}
\alias{page2table}
\title{A page2table Function}
\usage{
page2table(page.url, page.encoding = "UTF-8", export.file, css.selector)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
