% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.conf.R
\name{set.conf}
\alias{set.conf}
\title{A set.conf Function}
\usage{
set.conf(name, value, env.name = "env")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
