% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.filter.R
\name{signal.filter}
\alias{signal.filter}
\title{A m_std.data Function}
\usage{
signal.filter(signal, max = 0, co.indicator = NULL)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
