% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.CumulativeRate.R
\name{stock.CumulativeRate}
\alias{stock.CumulativeRate}
\title{A CumulativeRate Function}
\usage{
stock.CumulativeRate(
  data,
  strategy_type = NULL,
  strategy_params = c(9, 3, 3),
  trade_trigger = c(85, 20, 80, 20)
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
