% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_data.R
\name{testing_data}
\alias{testing_data}
\title{A for testing Function}
\usage{
testing_data(stock_name = "^GSPC", period = "2014-01-01/2015-04-30")
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
