% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeBysignal.R
\name{tradeBysignal}
\alias{tradeBysignal}
\title{A tradeBysignal Function}
\usage{
tradeBysignal(signal, price, limit.up = 0, limit.dn = 0, testing = FALSE)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
