% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upbreak.R
\name{upbreak}
\alias{upbreak}
\title{A upbreak Function}
\usage{
upbreak(Line, RefLine)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
