% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_crawler.R
\name{web_crawler}
\alias{web_crawler}
\title{A web_crawler Function}
\usage{
web_crawler(
  destfile,
  url,
  sleep = c(10:30),
  extra.holiday.date,
  fake.header = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
