% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muti_backtest.R
\name{muti_backtest}
\alias{muti_backtest}
\title{A muti_backtestt Function}
\usage{
muti_backtest(
  default.backtest_num = c(50, 5),
  default.trading.straregy_type = c(2, 80, 20, 80, 20),
  default.ef_goal_return = 0.001
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to multi paste vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
